Create Rectangle.js
Complete the function in the editor. It has two parameters: a and b. It must return an object modeling a rectangle that has the following properties:

length: This value is equal to a.
width: This value is equal to b.
perimeter: This value is equal to 2 * (a + b).
area: This value is equal to a * b
Note: The names of the object’s properties must be spelled correctly to pass this challenge.

Input Format
The first line contains an integer denoting a.
The second line contains an integer denoting b.

Constraints
1 <= a, b<= 100
Output Format
Return a object that has the properties specified above. Locked code in the editor prints the returned object’s length, width, perimeter, and area to STDOUT.

Sample Input 0

4
5
Sample Output 0

4
5
18
20
Explanation 0

Given a length of a = 4 and a width of b = 5, the Rectangle object’s perimeter is 4 + 4 +5 + 5 = 18 and its area is 4*5 = 20.