Create day5TemplateLiterals.js
The code in the editor has a tagged template literal that passes the area and perimeter of a rectangle to a tag function named sides. Recall that the first argument of a tag function is an array of string literals from the template, and the subsequent values are the template’s respective expression values.

Complete the function in the editor so that it does the following:

s = P +- (P^2 – 16*4)^1/2 /4

Finds the initial values of s1 and s2 by plugging the area and perimeter values into the formula:where A is the rectangle’s area and P is its perimeter.
Creates an array consisting of s1 and s2 and sorts it in ascending order.
Returns the sorted array.
Input Format
The first line contains an integer denoting s1.
The second line contains an integer denoting s2.

Constraints
1 <= s1, s2 <= 100
Output Format
Return an array consisting of s1 and s2, sorted in ascending order.

Sample Input 0

10
14
Sample Output 0

10
14
Explanation 0

The locked code in the editor passes the following arrays to the tag function:

The value of literals is [ 'The area is: ', '.\nThe perimeter is: ', '.' ].
The value of expressions is [ 140, 48 ], where the first value denotes the rectangle’s area, A, and the second value denotes its perimeter, P.
When we plug those values into our formula, we get the following:

s1 = 14
s2 = 10

We then store these values in an array, [14, 10], sort the array, and return the sorted array, [10, 14], as our answer.